Extract the CLI spawn helper to @studio/common
Moves the fork/spawn mechanics behind `createCliRunner` in @studio/common/lib/cli-process so non-desktop hosts can fork the CLI the same way; the desktop's execute-command now delegates to it and re-exports the same surface. No behavior change. First slice of splitting the studio ui PoC (#3953).
